Implement ClientScreen UI
This commit introduces the UI for the ClientsScreen, including a search bar and a list of clients displaying their ID and name. It also updates the ClientModel data class and adds new data classes for remote API responses. The default starting route in AndroidNavigation has been changed to AppRoute.Clients.route.